Pass className props to component
